Add files via upload
The growing application of surveillance, autonomous systems, and medical imaging relies tremendously on real-time video and image processing. The paper considers that there are sets of challenges to meet, including computational efficiency, accuracy, and robustness. In the discussed paper, a novel multi-resolution deep learning framework is developed to overcome these challenges. The framework consists of adaptive filtering and normalization for data pre-processing techniques, wavelet transforms, and western GSP-based multi-scale feature extraction module combined with convolutional neural networks (CNNs), as well as Transformer and recurrent neural network (RNN) based decision-making techniques. The combination of Particle swarm optimization and bio-inspired algorithm serves the purpose of optimizing performance that enhances overall results. Critically evaluating the performance on different metrics, it has been shown that the framework developed performs excellently on common metrics speed, accuracy, and computing effects on diverse datasets than the previous models. Such advantage will make it possible for more applications of image, signal, and video processing to be available for practical usage in real-time, and in operational environments that are memory lagged. Codes are available  at